Testes Criados

+++ src/main/java/domain/Roteiro.java
@@ -4,9 +4,9 @@ package domain;
 public class Roteiro {
     
     private final String nome;
-    private final float distancia;
+    private final double distancia;
     
-    public Roteiro(String nome, float distancia) {
+    public Roteiro(String nome, double distancia) {
         this.nome = nome;
         this.distancia = distancia;
 
@@ -16,7 +16,7 @@ public class Roteiro {
         return nome;
     }
 
-    public float getDistancia() {
+    public double getDistancia() {
         return distancia;
     }
